requirements.txt
python-telegram-bot
python-dotenv
sqlalchemy
pydantic
ollama
pytest
fastapi
uvicorn
pytz
